# Implement total time limit for a hard stop