--- a/src/papilo/core/PresolveOptions.hpp
+++ b/src/papilo/core/PresolveOptions.hpp
@@ -112,6 +112,7 @@ struct PresolveOptions
 
    double tlim = std::numeric_limits<double>::max();
 
+   double totaltlim = std::numeric_limits<double>::max();
 
    bool verification_with_VeriPB = false;
 
@@ -182,6 +183,8 @@ struct PresolveOptions
                              "times the number of rows or columns are active",
                              compressfac, 0.0, 1.0 );
       paramSet.addParameter( "presolve.tlim", "time limit for presolve", tlim,
+                             0.0 );
+      paramSet.addParameter( "presolve.totaltlim", "total time limit for presolve including the applying the reductions", totaltlim,
                              0.0 );
       paramSet.addParameter( "presolve.minabscoeff",
                              "minimum absolute coefficient value allowed in "